Fix compilation warnings

+++ src/eventserver/BasicHTTPServer_ObjC.mm
@@ -92,8 +92,8 @@ public:
     bool Request(char *senderIP, unsigned short senderPort, string *method, string *path, string *version, map<string, string> *headers, char *body, int bodylen) {
 
         if (strlen(body) != bodylen) {
-            NSLog(@"[UPnP] ERROR: real bodylen is %d, received bodylen is %d \n Content: %s", bodylen, strlen(body), body);
-            bodylen = strlen(body);
+            NSLog(@"[UPnP] ERROR: real bodylen is %d, received bodylen is %zu \n Content: %s", bodylen, strlen(body), body);
+            bodylen = (int)strlen(body);
         }
 
         @autoreleasepool {
